// angular/projects/models/post.interface.ts
import { Category } from "./category.";
import { User } from "./user.interface";

export interface Post {
        id: number;
        title: string;
        content: string;
        createdOn: Date;
        modifiedOn: Date;
        slug: string;
        category: Category;
        user: User;
        status?: string;
        mainImageUrl?: string;
}